requirements.txt
openai>=1.0.0
slack-bolt>=1.18.0
jinja2>=3.1.0
python-dotenv>=1.0.0
requests>=2.31.0
subprocess32>=3.5.4
pathlib>=1.0.1